Refactor API into modular routers
Split main.py (1189 lines) into focused routers:
- projects.py - Project CRUD operations
- colors.py - Color extraction endpoints
- sessions.py - Session/Library/Export endpoints
- main.py - App config, health, and status (169 lines)

Benefits:
- Better separation of concerns
- Easier to test and maintain
- Clear ownership of endpoints
- Aligned with FastAPI best practices
