+++ src/copy_that/interfaces/api/projects.py
@@ -1,0 +1,190 @@
+"""
+Project Management Router
+"""
+
+from fastapi import APIRouter, Depends, HTTPException, status
+from sqlalchemy.ext.asyncio import AsyncSession
+from sqlalchemy import select
+from datetime import datetime, timezone
+
+from copy_that.infrastructure.database import get_db
+from copy_that.domain.models import Project
+from copy_that.interfaces.api.schemas import (
+    ProjectCreateRequest,
+    ProjectUpdateRequest,
+    ProjectResponse,
+)
+
+router = APIRouter(prefix="/api/v1/projects", tags=["projects"])
+
+
+@router.post("", response_model=ProjectResponse, status_code=201)
+async def create_project(
+    request: ProjectCreateRequest,
+    db: AsyncSession = Depends(get_db)
+):
+    """Create a new project
+
+    Args:
+        request: ProjectCreateRequest with name and optional description
+        db: Database session
+
+    Returns:
+        Created project with ID
+    """
+    project = Project(
+        name=request.name,
+        description=request.description
+    )
+    db.add(project)
+    await db.commit()
+    await db.refresh(project)
+
+    return ProjectResponse(
+        id=project.id,
+        name=project.name,
+        description=project.description,
+        created_at=project.created_at.isoformat(),
+        updated_at=project.updated_at.isoformat()
+    )
+
+
+@router.get("", response_model=list[ProjectResponse])
+async def list_projects(
+    db: AsyncSession = Depends(get_db),
+    limit: int = 100,
+    offset: int = 0
+):
+    """List all projects
+
+    Args:
+        db: Database session
+        limit: Maximum number of projects to return
+        offset: Number of projects to skip
+
+    Returns:
+        List of projects
+    """
+    result = await db.execute(
+        select(Project).order_by(Project.created_at.desc()).limit(limit).offset(offset)
+    )
+    projects = result.scalars().all()
+
+    return [
+        ProjectResponse(
+            id=p.id,
+            name=p.name,
+            description=p.description,
+            created_at=p.created_at.isoformat(),
+            updated_at=p.updated_at.isoformat()
+        )
+        for p in projects
+    ]
+
+
+@router.get("/{project_id}", response_model=ProjectResponse)
+async def get_project(
+    project_id: int,
+    db: AsyncSession = Depends(get_db)
+):
+    """Get a specific project
+
+    Args:
+        project_id: Project ID
+        db: Database session
+
+    Returns:
+        Project details
+
+    Raises:
+        HTTPException: If project not found
+    """
+    result = await db.execute(select(Project).where(Project.id == project_id))
+    project = result.scalar_one_or_none()
+    if not project:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Project {project_id} not found"
+        )
+
+    return ProjectResponse(
+        id=project.id,
+        name=project.name,
+        description=project.description,
+        created_at=project.created_at.isoformat(),
+        updated_at=project.updated_at.isoformat()
+    )
+
+
+@router.put("/{project_id}", response_model=ProjectResponse)
+async def update_project(
+    project_id: int,
+    request: ProjectUpdateRequest,
+    db: AsyncSession = Depends(get_db)
+):
+    """Update a project
+
+    Args:
+        project_id: Project ID
+        request: ProjectUpdateRequest with fields to update
+        db: Database session
+
+    Returns:
+        Updated project
+
+    Raises:
+        HTTPException: If project not found
+    """
+    result = await db.execute(select(Project).where(Project.id == project_id))
+    project = result.scalar_one_or_none()
+    if not project:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Project {project_id} not found"
+        )
+
+    # Update fields if provided
+    if request.name is not None:
+        project.name = request.name
+    if request.description is not None:
+        project.description = request.description
+
+    project.updated_at = datetime.now(timezone.utc)
+
+    db.add(project)
+    await db.commit()
+    await db.refresh(project)
+
+    return ProjectResponse(
+        id=project.id,
+        name=project.name,
+        description=project.description,
+        created_at=project.created_at.isoformat(),
+        updated_at=project.updated_at.isoformat()
+    )
+
+
+@router.delete("/{project_id}", status_code=204)
+async def delete_project(
+    project_id: int,
+    db: AsyncSession = Depends(get_db)
+):
+    """Delete a project
+
+    Args:
+        project_id: Project ID
+        db: Database session
+
+    Raises:
+        HTTPException: If project not found
+    """
+    result = await db.execute(select(Project).where(Project.id == project_id))
+    project = result.scalar_one_or_none()
+    if not project:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Project {project_id} not found"
+        )
+
+    await db.delete(project)
+    await db.commit()
